refactor(bigquery): use Jobs.Instrumentation instead of a BigQuery-specific module
Consolidates on review feedback. BigQuery sync is the same category of thing as GCS
upload — a periodic Oban job wrapped for success/failure counting — and GCS already uses
Jobs.Instrumentation, so a sixth instrumentation module for a sibling case was
inconsistent.

The three things that appeared to block reuse were all limitations of the
Jobs.Instrumentation API rather than properties of the domain, so they are fixed there:

  - record/4 is now public and merges an extra tags map over the standard three, so
    BigQuery passes %{table: ..., action: ...} and per-table alerting is a group-by
    rather than a string match on the job name.
  - track_exception/4 records only on raise. track/3 derives success from the return
    value, which BigQuery cannot use: its response handler reacts to NOT_FOUND,
    PERMISSION_DENIED and TIMEOUT and then returns normally, so track/3 would classify
    all three as :success and report green while the sync is dead.
  - The status type is widened to cover the statuses BigQuery records precisely.

One metric for all periodic jobs now: job_run_count, tagged job / status /
organization_id plus table / action for the sync. GCS's track/3 path is untouched.

Note the cardinality change — job_run_count gains table and action dimensions for
bigquery_sync rows (18+ tables x 3 actions x orgs). Adding tags does not break existing
queries, but any alert already built on job_run_count is worth a sanity check.
